// src/enums/AnimObjects3D.ts
export enum NumberPlanes {
  upper = 'Upper',
  full = 'Full',
  lower = 'Lower',
  octants = 'Octants',
}

export enum ComplexPlanes {
  upper = 'Upper',
  full = 'Full',
  lower = 'Lower',
  octants = 'Octants',
}

export enum Octants {
  I = 'I',
  II = 'II',
  III = 'III',
  IV = 'IV',
  V = 'V',
  VI = 'VI',
  VII = 'VII',
  VIII = 'VIII',
}